+++ Assets/Tools/ScriptsDataBase/databasetables/EmpathyMap.cs
@@ -6,10 +6,10 @@ public class Empathymap
 {
 
     [PrimaryKey]
-    public int id { get; set; }
+    public Int64 id { get; set; }
     public int percentage { get; set; }
     public string creationDate { get; set; }
-    public int projectId { get; set; }
+    public Int64 projectId { get; set; }
     public string lastUpdate { get; set; }
 
     public override string ToString()
