+++ Assets/Tools/ScriptsDataBase/databasetables/Mindmap.cs
@@ -6,13 +6,13 @@ public class Mindmap
 {
 
     [PrimaryKey]
-    public int id { get; set; }
+    public Int64 id { get; set; }
 
     public int percentage { get; set; }
 
     public string creationDate { get; set; }
 
-    public int storytellingId { get; set; }
+    public Int64 storytellingId { get; set; }
 
     public string image { get; set; }
