+++ Assets/Tools/ScriptsDataBase/databasetables/Problem.cs
@@ -5,13 +5,13 @@ using System;
 public class Problem  {
 
 	[PrimaryKey]
-	public int id { get; set; }
+	public Int64 id { get; set; }
 	public int percentage { get; set; }
 	public string creationDate { get; set; }
 
 	public string lastUpdate { get; set; }
 
-	public int projectId { get; set; }
+	public Int64 projectId { get; set; }
 	// public int id;
 	// public int percentage;
 	// public string creationDate;
