+++ Assets/Tools/ScriptsDataBase/databasetables/Project.cs
@@ -5,14 +5,14 @@ using System;
 public class Project  {
 
 	[PrimaryKey]
-	public int id { get; set; }
+	public Int64 id { get; set; }
 	public string name { get; set; }
 	public int percentage { get; set; }
 	public string creationDate { get; set; }
 
 	public string sectorName { get; set; }
 
-	public int groupId { get; set; }
+	public Int64 groupId { get; set; }
 	public string lastUpdate { get; set; }
 	// public int id;
 	// public string name;
