+++ Assets/Tools/ScriptsDataBase/databasetables/Section.cs
@@ -5,12 +5,12 @@ using System;
 public class Section  {
 
 	[PrimaryKey]
-	public int id { get; set; }
+	public Int64 id { get; set; }
 	public string name { get; set; }
 	public string creationDate { get; set; }
 
 	
-	public int mindmapId { get; set; }
+	public Int64 mindmapId { get; set; }
 
 	public bool isOptional {get; set; }
